Normalized whitespace from previous commit
Also changed the interface to maviVibrate; uses unsigned int instead of
long (should fix compilation warnings).

+++ mavi-vibration.hpp
@@ -16,6 +16,6 @@ enum MaviVibratorID
 	MAVI_VIB_R
 };
 
-void maviVibrate(MaviVibratorID vibrator, double force, long duration);
+void maviVibrate(MaviVibratorID vibrator, double force, unsigned int duration);
 
 #endif
